Fix : integer -> float 변경

+++ Assets/Scripts/System/Buff/BuffProvider.cs
@@ -14,7 +14,7 @@ public class BuffProvider : MonoBehaviour
 		SetBuffer();
 	}
 
-	public void SetBuff(UnitBase unit, int buffCode, int buffActivityTime = -1)
+	public void SetBuff(UnitBase unit, int buffCode, float buffActivityTime = -1f)
 	{
 		ProceedBuff(unit, buffCode, buffActivityTime);
 	}
@@ -24,7 +24,7 @@ public class BuffProvider : MonoBehaviour
 		return HoldBuffDic.ContainsKey(buffCode);
 	}
 
-	private void ProceedBuff(UnitBase unit, int buffCode, int buffActivityTime = -1)
+	private void ProceedBuff(UnitBase unit, int buffCode, float buffActivityTime = -1f)
 	{
 		// Object Pooling은 안정화가 된 이후, 수정할 예정
 		var hasBuff = HasBuff(buffCode);
